pyproject.toml
[project]
name = "ko-embed"
version = "0.1.0"
description = "Korean query embedding pipeline for change-point detection research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "FlagEmbedding>=1.2",
    "sentence-transformers>=3.0",
    "umap-learn>=0.5",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "pandas>=2.0",
    "tqdm>=4.0",
    "hydra-core>=1.3",
    "mteb>=2.8.4",
    "kagglehub[pandas-datasets]>=1.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
